backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
xmltodict==0.13.0
lxml==5.3.0
pandas==2.2.2
openpyxl==3.1.5
pdfminer.six==20240706
PyMuPDF==1.24.9
pytesseract==0.3.13
Pillow==10.4.0
python-docx==1.1.2
reportlab==4.2.2
jinja2==3.1.4
httpx==0.27.2
python-dotenv==1.0.1
orjson==3.10.7
